+++ TestWebsite/Dto/PermissionDto.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace TestWebsite.Dto
+{
+    public class PermissionDto
+    {
+        public string Name { get; set; }
+        public List<string> Args { get; set; }
+        public bool IsPermitted { get; set; }
+    }
+}