+++ permissions/jest.config.js
@@ -1,0 +1,6 @@
+﻿module.exports = {
+    setupFiles: ['./src/__tests__/__helpers__/jest.setup.ts'],
+
+    // ignore files that don't have test.ts or test.tsx extension (.e.g. __helpers__)
+    testMatch: ["**/__tests__/**/*.test.ts?(x)"]
+}