+++ test-website/src/Api/AppPermissionsApi.ts
@@ -1,0 +1,31 @@
+import { PermissionName, Identity } from 'Models'
+import { GlobalPermissions } from '_Redux/Auth/GlobalPermissions'
+import {
+    CancellablePromise,
+} from '@interface-technologies/iti-react'
+import { get } from './ApiUtil'
+import { GetPermissionsApiMethod, PermissionDto, convenientGetFactory } from '@interface-technologies/permissions'
+
+export type AppPermissionsQueryTuple =
+    | [PermissionName.CanViewOrders]
+    | [PermissionName.CanManageOrders]
+    | [PermissionName.CanManageCustomer, Identity]
+    | [PermissionName.CanManageVendor, Identity]
+    | [PermissionName.CanManageCustomerVendorMap, Identity, Identity]
+
+const apiMethod: GetPermissionsApiMethod = q => get<PermissionDto[]>('api/appPermissions/get', {
+    q,
+})
+const convenientGet = convenientGetFactory<AppPermissionsQueryTuple>(apiMethod)
+
+function getGlobalPermissions(): CancellablePromise<GlobalPermissions> {
+    return convenientGet({
+        canViewOrders: [PermissionName.CanViewOrders],
+        canManageOrders: [PermissionName.CanManageOrders],
+    })
+}
+
+export const appPermissionsApi = {
+    get: convenientGet,
+    getGlobalPermissions,
+}