+++ test-website/src/Models/PermissionName.ts
@@ -1,0 +1,7 @@
+export enum PermissionName {
+    CanViewOrders = 'CanViewOrders',
+    CanManageOrders = 'CanManageOrders',
+    CanManageCustomer = 'CanManageCustomer',
+    CanManageVendor = 'CanManageVendor',
+    CanManageCustomerVendorMap = 'CanManageCustomerVendorMap'
+}