+++ test-website/src/Models/TypewriterModels/EmailAddress.ts
@@ -1,6 +1,12 @@
-﻿import { ErrorDtoType } from '.'
+﻿
+
+
+
+import { 
+    ErrorDtoType
+} from '.'
 
 export const EmailAddressTypeName = 'EmailAddress'
-export interface EmailAddress {
-    value: string
+export interface EmailAddress  { 
+	value: string
 }