+++ test-website/src/Models/TypewriterModels/ErrorDto.ts
@@ -1,8 +1,14 @@
-﻿import { ErrorDtoType } from '.'
+﻿
+
+
+
+import { 
+    ErrorDtoType
+} from '.'
 
 export const ErrorDtoTypeName = 'ErrorDto'
-export interface ErrorDto {
-    type: ErrorDtoType
-    message: string
-    diagnosticInfo: string
+export interface ErrorDto  { 
+	type: ErrorDtoType
+	message: string
+	diagnosticInfo: string
 }