+++ test-website/src/Models/TypewriterModels/ProductDto.ts
@@ -1,8 +1,14 @@
-﻿import { ErrorDtoType } from '.'
+﻿
+
+
+
+import { 
+    ErrorDtoType
+} from '.'
 
 export const ProductDtoTypeName = 'ProductDto'
-export interface ProductDto {
-    id: number
-    name: string
-    stock: number
+export interface ProductDto  { 
+	id: number
+	name: string
+	stock: number
 }