+++ test-website/src/Models/TypewriterModels/UserDto.ts
@@ -1,7 +1,13 @@
-﻿import { ErrorDtoType } from '.'
+﻿
+
+
+
+import { 
+    ErrorDtoType
+} from '.'
 
 export const UserDtoTypeName = 'UserDto'
-export interface UserDto {
-    id: number
-    name: string
+export interface UserDto  { 
+	id: number
+	name: string
 }