+++ test-website/src/Models/TypewriterModels/UserLogInDto.ts
@@ -1,7 +1,13 @@
-﻿import { ErrorDtoType } from '.'
+﻿
+
+
+
+import { 
+    ErrorDtoType
+} from '.'
 
 export const UserLogInDtoTypeName = 'UserLogInDto'
-export interface UserLogInDto {
-    accessToken: string
-    expiresUtc: Date
+export interface UserLogInDto  { 
+	accessToken: string
+	expiresUtc: Date
 }