+++ test-website/src/Models/index.ts
@@ -1,1 +1,3 @@
 export * from './TypewriterModels'
+export * from './PermissionName'
+export * from './Identity'