+++ test-website/src/Pages/Home/Index.tsx
@@ -69,6 +69,9 @@ export default class Page extends React.Component<PageProps, PageState> {
                     <li>
                         <Link to="/test/spamOnReady">Spam onReady</Link>
                     </li>
+                    <li>
+                        <Link to="/test/permissions">@interface-technologies/permissions test</Link>
+                    </li>
                 </ul>
             </div>
         )