+++ test-website/src/Pages/Test/TestRoutes.tsx
@@ -15,6 +15,7 @@ const TabManager = React.lazy(() => import('./TabManager/TabManager'))
 const UrlSearchParam = React.lazy(() => import('./UrlSearchParam'))
 const Hooks = React.lazy(() => import('./Hooks'))
 const SpamOnReady = React.lazy(() => import('./SpamOnReady'))
+const Permissions = React.lazy(() => import('./Permissions'))
 
 export const paths = {
     routeParam: '/test/routeParam/:number',
@@ -36,5 +37,6 @@ export function getTestRoutes(props: RoutesProps) {
         protectedRoute('/test/urlSearchParam', ppp(UrlSearchParam)),
         protectedRoute('/test/hooks', ppp(Hooks)),
         protectedRoute('/test/spamOnReady', ppp(SpamOnReady)),
+        protectedRoute('/test/permissions', ppp(Permissions)),
     ]
 }