+++ test-website/src/_Redux/Auth/GlobalPermissions.ts
@@ -1,0 +1,4 @@
+export interface GlobalPermissions {
+    canViewOrders: boolean
+    canManageOrders: boolean
+}